refactor Skills and Discipline select to use a common GenericSelect component
